Set ffmpeg process priority

+++ Xabe.FFmpeg/FFmpegWrapper.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.Win32.SafeHandles;
 using Xabe.FFmpeg.Events;
 using Xabe.FFmpeg.Exceptions;
 
@@ -18,7 +19,8 @@ namespace Xabe.FFmpeg
     /// </summary>
     internal class FFmpegWrapper : FFmpeg
     {
-        private const string TimeFormatRegex = @"\w\w:\w\w:\w\w";
+        private const string TimeFormatPattern = @"\w\w:\w\w:\w\w";
+        private static readonly Regex s_timeFormatRegex = new Regex(TimeFormatPattern, RegexOptions.Compiled);
         private List<string> _outputLog;
         private TimeSpan _totalTime;
         private bool _wasKilled = false;
@@ -33,14 +35,14 @@ namespace Xabe.FFmpeg
         /// </summary>
         internal event DataReceivedEventHandler OnDataReceived;
 
-        internal Task<bool> RunProcess(string args, CancellationToken cancellationToken)
+        internal Task<bool> RunProcess(
+            string args,
+            CancellationToken cancellationToken)
         {
-            return Task.Run(() =>
+            return Task.Factory.StartNew(() =>
             {
                 _outputLog = new List<string>();
-
-                var process = RunProcess(args, FFmpegPath, true, true, true);
-
+                var process = RunProcess(args, FFmpegPath, Priority, true, true, true);
                 using (process)
                 {
                     process.ErrorDataReceived += (sender, e) => ProcessOutputData(e, args);
@@ -49,25 +51,49 @@ namespace Xabe.FFmpeg
                     // https://github.com/Microsoft/vs-threading/blob/master/doc/analyzers/VSTHRD101.md
                     var ctr = cancellationToken.Register(() =>
                     {
-                        process.StandardInput.Write("q");
-                        Task.Delay(1000 * 5).ConfigureAwait(false).GetAwaiter().GetResult();
-
-                        try
+                        if (Environment.OSVersion.Platform != PlatformID.Win32NT)
                         {
-                            if (!process.HasExited)
+                            try
                             {
-                                _wasKilled = true;
-                                process.Kill();
+                                process.StandardInput.Write("q");
+                                Task.Delay(1000 * 5).ConfigureAwait(false).GetAwaiter().GetResult();
                             }
-                        }
-                        catch (InvalidOperationException)
-                        {
+                            catch (InvalidOperationException)
+                            {
+                            }
+                            finally
+                            {
+                                if (!process.HasExited)
+                                {
+                                    process.CloseMainWindow();
+                                    process.Kill();
+                                    _wasKilled = true;
+                                }
+                            }
                         }
                     });
 
                     using (ctr)
                     {
-                        process.WaitForExit();
+                        using (var processEnded = new ManualResetEvent(false))
+                        {
+                            processEnded.SetSafeWaitHandle(new SafeWaitHandle(process.Handle, false));
+                            int index = WaitHandle.WaitAny(new[] { processEnded, cancellationToken.WaitHandle });
+
+                            // If the signal came from the caller cancellation token close the window
+                            if (index == 1
+                                && !process.HasExited)
+                            {
+                                process.CloseMainWindow();
+                                process.Kill();
+                                _wasKilled = true;
+                            }
+                            else if (index == 0 && !process.HasExited)
+                            {
+                                // Workaround for linux: https://github.com/dotnet/corefx/issues/35544
+                                process.WaitForExit();
+                            }
+                        }
 
                         cancellationToken.ThrowIfCancellationRequested();
                         if (_wasKilled)
@@ -98,7 +124,10 @@ namespace Xabe.FFmpeg
                 }
 
                 return true;
-            });
+            },
+            cancellationToken,
+            TaskCreationOptions.LongRunning,
+            TaskScheduler.Default);
         }
 
         private void ProcessOutputData(DataReceivedEventArgs e, string args)
@@ -127,14 +156,13 @@ namespace Xabe.FFmpeg
                 return;
             }
 
-            var regex = new Regex(TimeFormatRegex);
             if (e.Data.Contains("Duration"))
             {
-                GetDuration(e, regex, args);
+                GetDuration(e, s_timeFormatRegex, args);
             }
             else if (e.Data.Contains("size"))
             {
-                Match match = regex.Match(e.Data);
+                Match match = s_timeFormatRegex.Match(e.Data);
                 if (match.Success)
                 {
                     OnProgress(this, new ConversionProgressEventArgs(TimeSpan.Parse(match.Value), _totalTime));
